requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=2.12=openblas
bzip2=1.0.8=h516909a_0
ca-certificates=2019.5.15=1
cairo=1.16.0=h18b612c_1001
certifi=2019.6.16=py36_1
cycler=0.10.0=py36_0
dbus=1.13.6=he372182_0
expat=2.2.5=he1b5a44_1003
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.0=he983fc9_1
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=h6f030ca_1002
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.5=nompi_h3c11f04_1103
icu=58.2=h9c2bf20_1
jasper=1.900.1=h07fcdf6_1006
jpeg=9c=h14c3975_1001
kiwisolver=1.1.0=py36he6710b0_0
lame=3.100=h14c3975_1001
libblas=3.8.0=12_openblas
libcblas=3.8.0=12_openblas
libclang=8.0.1=hc9558a2_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=12_openblas
liblapacke=3.8.0=12_openblas
libllvm8=8.0.1=hc9558a2_0
libopenblas=0.3.7=h6e990d7_1
libopus=1.3=h7b6447c_0
libpng=1.6.37=hed695b0_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h57b8799_1003
libuuid=2.32.1=h14c3975_1000
libvpx=1.7.0=h439df22_0
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxkbcommon=0.8.4=h516909a_0
libxml2=2.9.9=hea5a465_1
lz4-c=1.8.3=he1b5a44_1001
matplotlib=3.1.1=py36h5429711_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
nspr=4.20=hf484d3e_1000
nss=3.46=he751ad9_0
numpy=1.16.4=py36h99e49ec_0
numpy-base=1.16.4=py36h2f8d375_0
olefile=0.46=py36_0
opencv=4.1.1=py36ha799480_1
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1c=h7b6447c_1
pcre=8.41=hf484d3e_1003
pillow=6.1.0=py36h34e0f95_0
pip=19.2.2=py36_0
pixman=0.38.0=h516909a_1003
pthread-stubs=0.4=h14c3975_1001
pyparsing=2.4.2=py_0
pyqt=5.9.2=py36h751905a_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py36_0
pytz=2019.2=py_0
qt=5.9.7=h52cfd70_2
readline=7.0=h7b6447c_5
setuptools=41.0.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.29.0=h7b6447c_0
tk=8.6.8=hbc83047_0
tornado=6.0.3=py36h7b6447c_0
wheel=0.33.4=py36_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.8=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.4.0=h3b9ef0a_0
